Move config to separate file.
